Move is_overiden to deprecation_utils.py
